requirements.txt
fastapi>=0.115.0,<1.0.0
httpx>=0.28.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
uvicorn>=0.34.0,<1.0.0
